+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2024-02-23
+
+### Added
+
+- Added a configuration file for scoreboards
+
+### Misc
+
+- Version change triggered by a protocol change on rpcs3's side
+
+
 ## [1.1.0] - 2024-02-04
 
 ### Added